backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
anthropic
pydantic-settings
python-dotenv
httpx
bcrypt
docling
python-multipart
